split gl markers into gl2d and gl3d lists
- so that gl3d list remains the same until
  we spend more time tweaking each symbol's positioning in gl3d
