+++ src/constants/gl3d_markers.js
@@ -1,0 +1,21 @@
+/**
+* Copyright 2012-2017, Plotly, Inc.
+* All rights reserved.
+*
+* This source code is licensed under the MIT license found in the
+* LICENSE file in the root directory of this source tree.
+*/
+
+
+'use strict';
+
+module.exports = {
+    circle: '●',
+    'circle-open': '○',
+    square: '■',
+    'square-open': '□',
+    diamond: '◆',
+    'diamond-open': '◇',
+    cross: '+',
+    x: '❌'
+};
